+++ test/test-data-server.ts
@@ -3,6 +3,9 @@
 // In case of a fork or migration, these are all the information you will need to update/recreate.
 // ============================================================================================== //
 export const DATA_SERVER = {
+  get defects() {
+    return DEFECTS;
+  },
   get preconditions() {
     return PRECONDITIONS;
   },
@@ -83,6 +86,9 @@ const TESTS = {
       },
     ],
   },
+  immutableCucumber: {
+    key: "XCNODE-36",
+  },
   immutableDatadriven: {
     dataset: ["name,age", "Jane,42", "John,33", ""],
     key: "XCNODE-1",
@@ -91,6 +97,14 @@ const TESTS = {
   updateTestSteps: {
     v1: { key: "XCNODE-19", steps: [{ id: 3533 }] },
     v2: { key: "XCNODE-18", steps: [{ id: 3532 }] },
+  },
+} as const;
+// ============================================================================================== //
+// DEFECTS
+// ============================================================================================== //
+const DEFECTS = {
+  immutable: {
+    key: "XCNODE-33",
   },
 } as const;
 // ============================================================================================== //
@@ -107,7 +121,32 @@ const TEST_EXECUTIONS = {
   immutable: {
     key: "XCNODE-2",
     testEnvironments: [{ name: "environmentwith,comma" }, { name: "firefox" }],
-    tests: [{ ...TESTS.immutableDatadriven, testRunId: 12466 }, { ...TESTS.immutable }],
+    tests: [
+      {
+        ...TESTS.immutableDatadriven,
+        comment: { raw: "no errors", rendered: "<p>no errors</p>" },
+        status: "PASS",
+        testRunId: 12466,
+      },
+      { ...TESTS.immutable, defects: [DEFECTS.immutable], status: "FAIL", testRunId: 12805 },
+    ],
+  },
+  immutableCucumber: {
+    key: "XCNODE-37",
+    tests: [
+      {
+        ...TESTS.immutableCucumber,
+        examples: [
+          { id: 10043 },
+          { id: 10044, values: ["20", "30", "add", "50"] },
+          { id: 10045, values: ["2", "5", "add", "7"] },
+          { id: 10046, values: ["0", "40", "add", "40"] },
+          { id: 10047, values: ["4", "50", "add", "54"] },
+          { id: 10048, values: ["5", "50", "add", "55"] },
+        ],
+        testRunId: 13255,
+      },
+    ],
   },
   importXray: {
     v1: { key: "XCNODE-11", tests: [{ ...TESTS.immutableDatadriven }] },
@@ -119,6 +158,17 @@ const TEST_EXECUTIONS = {
   },
   removeTests: {
     key: "XCNODE-31",
+  },
+  updateCucumberExamples: {
+    key: "XCNODE-38",
+    tests: [
+      {
+        ...TESTS.immutableCucumber,
+        initialStatuses: ["PASS"],
+        testRunId: 13256,
+        updatedStatuses: ["FAIL"],
+      },
+    ],
   },
   updateTestRun: {
     v1: {
@@ -136,6 +186,40 @@ const TEST_EXECUTIONS = {
       ],
     },
   },
+  updateTestRunComments: {
+    key: "XCNODE-35",
+    tests: [
+      {
+        ...TESTS.immutable,
+        initialComment: { raw: "initial comment", rendered: "<p>initial comment</p>" },
+        testRunId: 13253,
+        updatedComment: { raw: "updated comment", rendered: "<p>updated comment</p>" },
+      },
+    ],
+  },
+  updateTestRunDefects: {
+    key: "XCNODE-34",
+    testRuns: {
+      addDefects: {
+        defects: [DEFECTS.immutable],
+        status: "FAIL",
+        test: TESTS.immutable,
+        testRunId: 13251,
+      },
+      removeDefects: {
+        defects: [DEFECTS.immutable],
+        status: "FAIL",
+        test: TESTS.immutableDatadriven,
+        testRunId: 13252,
+      },
+    },
+  },
+  updateTestRunStatuses: {
+    key: "XCNODE-32",
+    tests: [
+      { ...TESTS.immutable, initialStatus: "TODO", testRunId: 13250, updatedStatus: "EXECUTING" },
+    ],
+  },
 } as const;
 // ============================================================================================== //
 // TEST PLANS
@@ -150,7 +234,7 @@ const TEST_PLANS = {
   immutable: {
     key: "XCNODE-7",
     testExecutions: [TEST_EXECUTIONS.immutable, TEST_EXECUTIONS.addAttachments.v2],
-    tests: [TESTS.immutableDatadriven, TESTS.immutable],
+    tests: [TESTS.immutableDatadriven, TESTS.immutable, TESTS.immutableCucumber],
   },
   removeTestExecutions: {
     key: "XCNODE-29",