+++ CHANGELOG.md
@@ -14,6 +14,16 @@
 - Added `DELETE /rest/raven/1.0/api/testplan/{testPlanKey}/testexecution/{testExecKey}`
 - Added `POST   /rest/raven/1.0/api/testexec/{testExecKey}/test`
 - Added `DELETE /rest/raven/1.0/api/testexec/{testExecKey}/test/{testKey}`
+- Added `GET    /rest/raven/1.0/api/testrun/{id}/status`
+- Added `PUT    /rest/raven/1.0/api/testrun/{id}/status`
+- Added `GET    /rest/raven/1.0/api/testrun/{id}/defect`
+- Added `POST   /rest/raven/1.0/api/testrun/{id}/defect`
+- Added `DELETE /rest/raven/1.0/api/testrun/{id}/defect/{issueIdOrKey}`
+- Added `GET    /rest/raven/1.0/api/testrun/{id}/comment`
+- Added `PUT    /rest/raven/1.0/api/testrun/{id}/comment`
+- Added `GET    /rest/raven/1.0/api/testrun/{id}/example`
+- Added `GET    /rest/raven/1.0/api/testrun/{id}/example/{exampleIndex}`
+- Added `PUT    /rest/raven/1.0/api/testrun/{id}/example/{exampleIndex}`
 
 # `4.0.0`